registry/recipes/docling-pipeline/app/requirements.txt
docling>=2.70.0,<3.0.0
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
python-multipart>=0.0.12,<1.0.0
PyYAML>=6.0.0,<7.0.0
